+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- "Audio/Midi &Settings" menu item has been renamed to "Settings..." and has been moved to the File menu
 - Added capability of setting the UI language https://github.com/oleg68/GrandOrgue/issues/45
 - Fixed formatting of .md files
 # 0.3.1.2341-6.os (2021-07-07)